requirements.txt
vnstock3>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ta>=0.11.0
yfinance>=0.2.0
plotly>=5.18.0
streamlit>=1.29.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
